Move kernels from src to cpu_solver directory
