+++ reactions-diffusion/biofvm/CMakeLists.txt
@@ -15,7 +15,7 @@ target_sources(
 target_link_libraries(physicore.reactions-diffusion.biofvm
                       PUBLIC physicore::common)
 
-add_subdirectory(src/kernels/cpu_solver)
+add_subdirectory(kernels/cpu_solver)
 
 if(PHYSICORE_BUILD_TESTS)
   find_package(GTest REQUIRED)
