+++ reactions-diffusion/biofvm/kernels/cpu_solver/include/bulk_solver.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../../../include/microenvironment.h"
+#include "../../../include/microenvironment.h"
 #include "diffusion_solver.h"
 
 /*
